+++ packages/multichain-api-middleware/src/index.test.ts
@@ -10,6 +10,7 @@ describe('@metamask/multichain-api-middleware', () => {
         "multichainMethodCallValidatorMiddleware",
         "MultichainMiddlewareManager",
         "MultichainSubscriptionManager",
+        "MultichainApiNotifications",
       ]
     `);
   });