+++ packages/multichain-api-middleware/src/index.ts
@@ -5,3 +5,4 @@ export { walletRevokeSession } from './handlers/wallet-revokeSession';
 export { multichainMethodCallValidatorMiddleware } from './middlewares/multichainMethodCallValidatorMiddleware';
 export { MultichainMiddlewareManager } from './middlewares/MultichainMiddlewareManager';
 export { MultichainSubscriptionManager } from './middlewares/MultichainSubscriptionManager';
+export { MultichainApiNotifications } from './handlers/types';